pyproject.toml
[project]
name = "meu-projeto"
version = "0.1.0"
description = ""
authors = [
    {name = "CarlosSenra",email = "crsbsenra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "numpy (>=2.3.1,<3.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "uvicorn (==0.32.1)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pynest-api (>=0.4.0,<0.5.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]

[tool.poetry]
packages = [{include = "meu_projeto", from = "src"}]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
mypy = "^1.16.1"
pre-commit = "^4.2.0"
pytest = "^8.4.1"
black = ">=24.10.0,<25.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
extend-ignore = ["E203", "W503", "E501"]
max-line-length = 88